requirements.txt
# requirements.txt
fastapi
uvicorn
speechrecognition
pyttsx3
pyaudio
python-dotenv
requests
wikipedia
secure-smtplib
